Move shared module to shared
