+++ apps/web/src/app/modules/organizations/create/organization-create.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../../shared.module";
+import { SharedModule } from "../../../shared";
 
 import { OrganizationInformationComponent } from "./organization-information.component";
