+++ apps/web/src/app/modules/organizations/manage/organization-manage.module.ts
@@ -1,7 +1,7 @@
 import { ScrollingModule } from "@angular/cdk/scrolling";
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../../shared.module";
+import { SharedModule } from "../../../shared";
 
 import { EntityUsersComponent } from "./entity-users.component";
